requirements.txt
Pywin32
PyQt5
faster-whisper
Vosk
Pyaudio
Openai